services/ai/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
motor==3.5.0
pydantic==2.9.0
PyJWT==2.9.0
python-multipart==0.0.9
httpx==0.27.0
pytesseract==0.3.13
Pillow==10.4.0
pyzbar==0.1.9
opencv-python-headless==4.10.0.84
numpy==1.26.4
requests>=2.31.0
